+++ PhantomEngine/RHI/OpenGL/OpenGLGraphicsManager.cpp
@@ -259,7 +259,7 @@ namespace Phantom {
 		glFlush();
 	}
 
-	void OpenGLGraphicsManager::resize(float width, float height)
+	void OpenGLGraphicsManager::resize(int32_t width, int32_t height)
 	{
 		glViewport(0, 0, width, height);
 	}
